@@ -11,7 +11,7 @@
     <parent>
         <groupId>org.springframework.boot</groupId>
         <artifactId>spring-boot-starter-parent</artifactId>
-        <version>2.1.7.RELEASE</version>
+        <version>2.4.5</version>
     </parent>
 
     <properties>
@@ -24,7 +24,7 @@
         <shedlock.version>2.2.0</shedlock.version>
         <logstash-logback-encoder.version>6.1</logstash-logback-encoder.version>
         <okhttp.version>3.6.0</okhttp.version>
-        <java.version>12</java.version>
+        <java.version>14</java.version>
     </properties>
 
 
