@@ -5,13 +5,14 @@ import no.nav.pam.annonsemottak.Application;
 import no.nav.pam.annonsemottak.PathDefinition;
 import no.nav.pam.annonsemottak.app.config.TestConfig;
 import no.nav.pam.annonsemottak.receivers.amedia.AmediaResponseMapperTest;
-import no.nav.pam.annonsemottak.receivers.amedia.AmediaService;
 import no.nav.pam.annonsemottak.stilling.StillingRepository;
 import org.assertj.core.api.SoftAssertions;
-import org.junit.Before;
-import org.junit.Rule;
-import org.junit.Test;
-import org.junit.runner.RunWith;
+
+import org.junit.jupiter.api.AfterAll;
+import org.junit.jupiter.api.BeforeAll;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
 import org.springframework.boot.test.context.SpringBootTest;
@@ -19,7 +20,8 @@ import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.http.MediaType;
 import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.context.ContextConfiguration;
-import org.springframework.test.context.junit4.SpringRunner;
+import org.springframework.test.context.junit.jupiter.SpringExtension;
+
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.MvcResult;
 
@@ -31,7 +33,7 @@ import static com.github.tomakehurst.wiremock.client.WireMock.*;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
-@RunWith(SpringRunner.class)
+@ExtendWith(SpringExtension.class)
 @SpringBootTest(classes = Application.class)
 @MockBean(StillingRepository.class)
 @AutoConfigureMockMvc
@@ -46,14 +48,14 @@ public class AmediaApiTest {
         System.setProperty("DEXI_API_PASSWORD", "");
     }
 
-    @Rule
-    public WireMockRule wireMockRule = new WireMockRule(7010);
+    public static WireMockRule wireMockRule;
 
     @Autowired
     protected MockMvc mvc;
 
-    @Before
-    public void initStubs() {
+    @BeforeAll
+    public static void initStubs() {
+        wireMockRule = new WireMockRule(7010);
         wireMockRule.stubFor(get(urlMatching(
                 ".*?all"))
                 .willReturn(aResponse().withStatus(200)
@@ -65,13 +67,19 @@ public class AmediaApiTest {
                 .willReturn(aResponse().withStatus(200)
                         .withHeader("Content-Type", "application/json")
                         .withBody(getMockResponse("enkeltResultat"))));
+        wireMockRule.start();
     }
 
-    private String getMockResponse(String fil) {
+    private static String getMockResponse(String fil) {
         return new BufferedReader(
                 new InputStreamReader(AmediaResponseMapperTest.class.getClassLoader()
                         .getResourceAsStream("amedia.io.samples/" + fil + ".json")))
                 .lines().collect(Collectors.joining("\n"));
+    }
+
+    @AfterAll
+    public static void stopWireMock() {
+        wireMockRule.stop();
     }
 
     /*
