@@ -3,7 +3,7 @@ package no.nav.pam.annonsemottak.receivers.xmlstilling;
 import no.nav.pam.annonsemottak.receivers.xmlstilling.Stillinger.Gruppe;
 import no.nav.pam.annonsemottak.stilling.Stilling;
 import no.nav.pam.annonsemottak.stilling.StillingTestdataBuilder;
-import org.junit.Test;
+import org.junit.jupiter.api.Test;
 
 import java.time.LocalDateTime;
 import java.util.Arrays;
