@@ -1,6 +1,6 @@
 package no.nav.pam.annonsemottak.rest.payloads;
 
-import org.junit.Test;
+import org.junit.jupiter.api.Test;
 
 import java.util.Map;
 
