@@ -1,7 +1,8 @@
 package no.nav.pam.annonsemottak.stilling;
 
-import org.junit.Assert;
-import org.junit.Test;
+
+import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.Test;
 
 import static no.nav.pam.annonsemottak.stilling.StillingTestdataBuilder.enkelStilling;
 
@@ -10,15 +11,15 @@ public class AnnonseStatusTest {
     @Test
     public void stop() {
         Stilling nyStilling = enkelStilling().build();
-        Assert.assertEquals(nyStilling.getAnnonseStatus(), AnnonseStatus.AKTIV);
+        Assertions.assertEquals(nyStilling.getAnnonseStatus(), AnnonseStatus.AKTIV);
         nyStilling.stop();
-        Assert.assertEquals(nyStilling.getAnnonseStatus(), AnnonseStatus.STOPPET);
+        Assertions.assertEquals(nyStilling.getAnnonseStatus(), AnnonseStatus.STOPPET);
     }
 
     @Test
     public void deactivate() {
         Stilling nyStilling = enkelStilling().build();
         nyStilling.deactivate();
-        Assert.assertEquals(nyStilling.getAnnonseStatus(), AnnonseStatus.INAKTIV);
+        Assertions.assertEquals(nyStilling.getAnnonseStatus(), AnnonseStatus.INAKTIV);
     }
 }
