@@ -1,6 +1,7 @@
 package no.nav.pam.annonsemottak.stilling;
 
-import org.junit.Test;
+
+import org.junit.jupiter.api.Test;
 
 import static org.hamcrest.CoreMatchers.equalTo;
 import static org.hamcrest.CoreMatchers.is;
