@@ -1,6 +1,6 @@
 package no.nav.pam.annonsemottak.stilling;
 
-import org.junit.Test;
+import org.junit.jupiter.api.Test;
 import org.springframework.data.domain.Sort;
 
 import static org.hamcrest.CoreMatchers.equalTo;
@@ -12,31 +12,31 @@ public class SorteringTest {
     @Test
     public void sortering_skal_støtte_mottattdato() {
         Sortering sorteringEtterMottattdatoAscending = Sortering.valueOf(Sortering.OrderBy.nullSafeValueOf("Mottattdato"), null);
-        assertThat(sorteringEtterMottattdatoAscending.asSort(), is(equalTo(new Sort(Sort.Direction.ASC, "created"))));
+        assertThat(sorteringEtterMottattdatoAscending.asSort(), is(equalTo(Sort.by(Sort.Direction.ASC, "created"))));
     }
 
     @Test
     public void sortering_skal_støtte_arbeidsgiver() {
         Sortering sorteringEtterMottattdatoAscending = Sortering.valueOf(Sortering.OrderBy.nullSafeValueOf("Arbeidsgiver"), null);
-        assertThat(sorteringEtterMottattdatoAscending.asSort(), is(equalTo(new Sort(Sort.Direction.ASC, "employer"))));
+        assertThat(sorteringEtterMottattdatoAscending.asSort(), is(equalTo(Sort.by(Sort.Direction.ASC, "employer"))));
     }
 
     @Test
     public void sortering_skal_støtte_arbeidssted() {
         Sortering sorteringEtterMottattdatoAscending = Sortering.valueOf(Sortering.OrderBy.nullSafeValueOf("Sted"), null);
-        assertThat(sorteringEtterMottattdatoAscending.asSort(), is(equalTo(new Sort(Sort.Direction.ASC, "place"))));
+        assertThat(sorteringEtterMottattdatoAscending.asSort(), is(equalTo(Sort.by(Sort.Direction.ASC, "place"))));
     }
 
     @Test
     public void sortering_skal_støtte_stillingstittel() {
         Sortering sorteringEtterMottattdatoAscending = Sortering.valueOf(Sortering.OrderBy.nullSafeValueOf("Tittel"), null);
-        assertThat(sorteringEtterMottattdatoAscending.asSort(), is(equalTo(new Sort(Sort.Direction.ASC, "title"))));
+        assertThat(sorteringEtterMottattdatoAscending.asSort(), is(equalTo(Sort.by(Sort.Direction.ASC, "title"))));
     }
 
     @Test
     public void asSort_skal_lage_samme_objekt_som_bruk_av_sort_sin_constructor_direkte() {
         Sort sort = Sortering.valueOf(Sortering.OrderBy.MOTTATTDATO, null).asSort();
-        Sort sort2 = new Sort(Sort.Direction.ASC, "created");
+        Sort sort2 = Sort.by(Sort.Direction.ASC, "created");
 
         assertThat(sort, is(equalTo(sort2)));
     }
